vesper2.0/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vesper"
version = "0.1.0"
description = "Habitat 3.0 + IoT Interactive Simulation Testbed"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Vesper Team"}
]
keywords = ["habitat", "simulation", "iot", "robotics", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.21.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
habitat = [
    "habitat-sim",
    "habitat-lab",
]
llm = [
    "openai>=1.0",
    "tiktoken",
]
network = [
    "aiohttp>=3.8",
    "websockets>=10.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "pytest-asyncio",
    "black",
    "ruff",
    "mypy",
]
# For Windows/platforms without habitat-sim support
no-habitat = [
    "vesper[llm,network,dev]",
]
all = [
    "vesper[habitat,llm,network,dev]",
]

[project.urls]
Homepage = "https://github.com/vesper-sim/vesper"
Documentation = "https://github.com/vesper-sim/vesper#readme"
Repository = "https://github.com/vesper-sim/vesper"

[tool.setuptools.packages.find]
where = ["."]
include = ["vesper*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
